Windows: Fixed stat warnings.

+++ 3rdparty/libprocess/3rdparty/stout/include/stout/os/windows/stat.hpp
@@ -126,7 +126,7 @@ inline Try<long> mtime(const std::string& path)
     return ErrnoError("Error invoking stat for '" + path + "'");
   }
 
-  return s.st_mtime;
+  return (long)s.st_mtime;
 }
